model solution for Ex90
The solution used the same methods recommended by CoPilot but only used the if side of the branch and returns false in all cases that don't go down on either of the ifs.
Also, the last section in the code, this if __name__ == '__mani__' appears in the code. I believe this is for ensuring main() doesn't run when the function is imported and it looks familiar but I need to spend a little more time to revise what is its purpose and how to use it properly.